+++ Krowi_Databases/DbManager/DbManagerWPF/DataManager/AchievementDM.cs
@@ -93,7 +93,7 @@ namespace DbManagerWPF.DataManager
         {
             if (ID <= 0) throw new ArgumentOutOfRangeException(nameof(ID));
             if (location <= 0) throw new ArgumentOutOfRangeException(nameof(location));
-            if (points <= 0) throw new ArgumentOutOfRangeException(nameof(points));
+            if (points < 0) throw new ArgumentOutOfRangeException(nameof(points));
             _ = category ?? throw new ArgumentNullException(nameof(category));
 
             var sb = new StringBuilder();
@@ -286,6 +286,44 @@ namespace DbManagerWPF.DataManager
             cmd.Parameters.AddWithValue("@NewCategoryID", newParent.ID);
             cmd.Parameters.AddWithValue("@OldCategoryID", oldParent.ID);
             cmd.Parameters.AddWithValue("@AchievementID", achievement.ID);
+
+            cmd.ExecuteNonQuery();
+        }
+
+        public void UpdateAGT(Achievement achievement, string rewardText, int category_AGT_ID, int uiOrder, int iconFileID)
+        {
+            var cmd = connection.CreateCommand();
+            cmd.CommandText = @"INSERT OR REPLACE INTO Achievement_AGT VALUES (@ID, @Name, @Description, @RewardText, @FactionID, @Category_AGT_ID, @Points, @Flags, @UIOrder, @IconFileID, @CovenantID, @HIDE_INCOMPLETE, @TRACKING_FLAG,
+	                                CASE WHEN (SELECT COUNT(*) FROM Achievement_AGT WHERE ID = @ID) > 0 THEN
+		                                (SELECT DateAdded FROM Achievement_AGT WHERE ID = @ID) ELSE DATETIME('now', 'localtime')
+	                                END,
+	                                CASE WHEN ((SELECT Name FROM Achievement_AGT WHERE ID = @ID) = @Name)
+		                                        AND ((SELECT Description FROM Achievement_AGT WHERE ID = @ID) = @Description)
+		                                        AND ((SELECT RewardText FROM Achievement_AGT WHERE ID = @ID) = @RewardText)
+		                                        AND ((SELECT FactionID FROM Achievement_AGT WHERE ID = @ID) = @FactionID)
+		                                        AND ((SELECT Category_AGT_ID FROM Achievement_AGT WHERE ID = @ID) = @Category_AGT_ID)
+		                                        AND ((SELECT Points FROM Achievement_AGT WHERE ID = @ID) = @Points)
+		                                        AND ((SELECT Flags FROM Achievement_AGT WHERE ID = @ID) = @Flags)
+		                                        AND ((SELECT UIOrder FROM Achievement_AGT WHERE ID = @ID) = @UIOrder)
+		                                        AND ((SELECT IconFileID FROM Achievement_AGT WHERE ID = @ID) = @IconFileID)
+		                                        AND ((SELECT CovenantID FROM Achievement_AGT WHERE ID = @ID) = @CovenantID)
+		                                        AND ((SELECT HIDE_INCOMPLETE FROM Achievement_AGT WHERE ID = @ID) = @HIDE_INCOMPLETE)
+		                                        AND ((SELECT TRACKING_FLAG FROM Achievement_AGT WHERE ID = @ID) = @TRACKING_FLAG) THEN
+		                                (SELECT DateChanged FROM Achievement_AGT WHERE ID = @ID) ELSE DATETIME('now', 'localtime')
+	                                END);";
+            cmd.Parameters.AddWithValue("@ID", achievement.ID);
+            cmd.Parameters.AddWithValue("@Name", achievement.Name);
+            cmd.Parameters.AddWithValue("@Description", achievement.Description != null ? achievement.Description : DBNull.Value);
+            cmd.Parameters.AddWithValue("@RewardText", string.IsNullOrEmpty(rewardText) ? DBNull.Value : rewardText);
+            cmd.Parameters.AddWithValue("@FactionID", (int)achievement.Faction);
+            cmd.Parameters.AddWithValue("@Category_AGT_ID", category_AGT_ID);
+            cmd.Parameters.AddWithValue("@Points", achievement.Points);
+            cmd.Parameters.AddWithValue("@Flags", (int)achievement.Flags);
+            cmd.Parameters.AddWithValue("@UIOrder", uiOrder);
+            cmd.Parameters.AddWithValue("@IconFileID", iconFileID);
+            cmd.Parameters.AddWithValue("@CovenantID", (int)achievement.Covenant);
+            cmd.Parameters.AddWithValue("@HIDE_INCOMPLETE", achievement.Flags.HasFlag(AchievementFlags.HIDE_INCOMPLETE) ? 1 : 0);
+            cmd.Parameters.AddWithValue("@TRACKING_FLAG", achievement.Flags.HasFlag(AchievementFlags.TRACKING_FLAG) ? 1 : 0);
 
             cmd.ExecuteNonQuery();
         }