+++ Krowi_Databases/DbManager/DbManagerWPF/ViewModel/MainWindowViewModel.cs
@@ -26,6 +26,7 @@ namespace DbManagerWPF.ViewModel
 
             LoadCategoriesViewModel();
             LoadUIMapViewModel();
+            LoadImportViewModel();
         }
 
         public event PropertyChangedEventHandler PropertyChanged;