+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [9.0.5.24.0] - 
+## [9.1.0.24.0] - 
 ### Added
-- TBD
+- More zone data for multi-zone achievements for Shadowlands
 
 ### Changed
 - Backend data structure completely rewritten in preperation for new features
+- Coming in 9.1 merged into the other categories
+- Moved Covenant Campaign achievement from Shadowlands - Zones to Shadowlands - Covenant Sanctum
 
 ### Fixed
 - Northrend Explorer now shows for both factions
@@ -445,7 +447,7 @@ All notable changes to this project will be documented in this file.
 ### Removed
 - The need for almost all localization
 
-[9.0.5.24.0]: https://github.com/TheKrowi/Krowi_Addons/tree/master/Krowi_AchievementFilter
+[9.1.0.24.0]: https://github.com/TheKrowi/Krowi_Addons/tree/master/Krowi_AchievementFilter
 [9.0.5.23.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3326868
 [9.0.5.22.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3308327
 [9.0.5.21.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3302611